+++ mod-source-record-manager-server/src/main/java/org/folio/verticle/consumers/StoredRecordChunksKafkaHandler.java
@@ -4,6 +4,7 @@ import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
+import io.vertx.core.json.Json;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.dataimport.util.OkapiConnectionParams;
@@ -61,7 +62,7 @@ public class StoredRecordChunksKafkaHandler implements AsyncRecordHandler<String
     String correlationId = okapiConnectionParams.getHeaders().get("correlationId");
     String chunkNumber = okapiConnectionParams.getHeaders().get("chunkNumber");
 
-    Event event = new JsonObject(record.value()).mapTo(Event.class);
+    Event event = Json.decodeValue(record.value(), Event.class);
 
     if (!kafkaInternalCache.containsByKey(event.getId())) {
       try {